a fs_offset_t is a fs_offset_t is a fs_offset_t is a fs_offset_t and not a long
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@7829 d7cf8633-e32d-0410-b094-e92efae38249

+++ fs.c
@@ -2169,7 +2169,7 @@ int FS_Seek (qfile_t* file, fs_offset_t offset, int whence)
 		default:
 			return -1;
 	}
-	if (offset < 0 || offset > (long) file->real_length)
+	if (offset < 0 || offset > file->real_length)
 		return -1;
 
 	// If we have the data in our read buffer, we don't need to actually seek
